[Fix] changed chat_id and user_id to BigInteger (issues with supergroup)

+++ sql/domain/ChatEntity.py
@@ -2,7 +2,7 @@
 
 from datetime import datetime
 
-from sqlalchemy import Column, Integer, String, TIMESTAMP
+from sqlalchemy import Column, String, TIMESTAMP, BigInteger
 from sqlalchemy.orm import relationship
 
 from sql import Base
@@ -11,7 +11,7 @@ from sql import Base
 class Chat(Base):
     __tablename__ = 'chat'
 
-    chat_id = Column(Integer, primary_key=True, nullable=False, unique=True)
+    chat_id = Column(BigInteger, primary_key=True, nullable=False, unique=True)
     name = Column(String, nullable=False)
 
     created_at = Column(TIMESTAMP, nullable=False, default=datetime.now())
